Improved (corrected a missing case) of the 2nd axiom.
The axiom: If n columns/rows contain the entirety of n colorsets, the cells of all other colors within those n columns/rows can be crossed.

Earlier we used: rather than if len(holdings) != len(colors):
But instead we can use:
"for each color, is the number of others' holdings been subsets of this colors holdings is the same as the length of the holding".
This is more complete.

This resulted in a significant efficiency gain in puzzle 20250902, which benefits from this axiom been completely implemented.